got some better data!
ok so I reran the data with the current (heh) settings in
beamSearchWithError. we got  a few data points but nothing with the
other side of the mirror (the z axis went off the sensor) anyway have
fun there’s at least a relatively similar slope to the quadrant 1 data
run!